\name{coef.sbss}
\alias{coef.sbss}

\title{
Coef Method for an Object of Class 'sbss'
}

\description{
Extracts the estimated unmixing matrix of an object of class \code{'sbss'}.
}

\usage{
\method{coef}{sbss}(object, ...)
}

\arguments{
  \item{object}{
object of class \code{'sbss'}. Usually result of \code{\link{sbss}}.
}
  \item{\dots}{
further arguments to be passed to or from methods.
}
}

\value{
Returns the estimated unmixing matrix of an object of class \code{'sbss'} as a numeric matrix.
}

\seealso{
\code{\link{sbss}}
}
