\name{print.sbss}
\alias{print.sbss}

\title{
Print Method for an Object of Class 'sbss'
}

\description{
Prints the estimated unmixing matrix and the diagonalized local covariance matrices for an object of class \code{'sbss'}.
}

\usage{
\method{print}{sbss}(x, \dots)
}

\arguments{
  \item{x}{
object of class \code{'sbss'}. Usually result of \code{\link{sbss}}.
}
  \item{\dots}{
additional arguments for the method \code{print.listof}.
}
}

\seealso{
\code{\link{sbss}}
}

